+++ CHANGELOG.md
@@ -10,6 +10,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Make `Iban#REGEX` public (as part of #171).
 - Add `SwiftPatternCharacterRepresentation` to centralize information about SWIFT pattern character representations (as
   part of #171).
+- Add random IBAN generation for all the ISO 13616-compliant national IBAN formats (#153).
+- Make `BbanStructure` public (as part of #153). `BbanStructure` is an enum that holds countries BBAN structure as
+  listed in the [IBAN registry](https://www.iso13616.org).
 
 ### Changed